resolve the comments from pr for transition extractor
