+++ apps/builder/app/builder/features/style-panel/sections/transforms/transform-utils.test.ts
@@ -6,6 +6,7 @@ import {
   isTransformPanelPropertyUsed,
   updateRotateOrSkewPropertyValue,
   updateTransformTuplePropertyValue,
+  extractRotatePropertiesFromTransform,
 } from "./transform-utils";
 import type { StyleInfo } from "../../shared/style-info";
 import {
@@ -15,10 +16,7 @@ import {
   type StyleProperty,
   type StyleValue,
 } from "@webstudio-is/css-engine";
-import {
-  extractRotatePropertiesFromTransform,
-  parseCssValue,
-} from "@webstudio-is/css-data";
+import { parseCssValue } from "@webstudio-is/css-data";
 
 const initializeStyleInfo = () => {
   const currentStyle: StyleInfo = {};
@@ -263,5 +261,56 @@ describe("Transform utils CRUD operations", () => {
     expect(toValue(newValue)).toBe(
       "rotateX(10deg) rotateY(50deg) rotateZ(10deg) skewX(10deg) skewY(10deg)"
     );
+  });
+});
+
+describe("extractRotatePropertiesFromTransform", () => {
+  test("parses transform and returns undefined if no rotate values exists", () => {
+    expect(
+      extractRotatePropertiesFromTransform(
+        parseCssValue("transform", "scale(1.5)")
+      )
+    ).toEqual({
+      rotateX: undefined,
+      rotateY: undefined,
+      rotateZ: undefined,
+    });
+  });
+
+  test("parses transform and returns rotate values", () => {
+    expect(
+      extractRotatePropertiesFromTransform(
+        parseCssValue("transform", "rotateX(0deg) rotateY(10deg) scale(1.5)")
+      )
+    ).toEqual({
+      rotateX: {
+        type: "function",
+        args: {
+          type: "layers",
+          value: [
+            {
+              type: "unit",
+              unit: "deg",
+              value: 0,
+            },
+          ],
+        },
+        name: "rotateX",
+      },
+      rotateY: {
+        type: "function",
+        args: {
+          type: "layers",
+          value: [
+            {
+              type: "unit",
+              unit: "deg",
+              value: 10,
+            },
+          ],
+        },
+        name: "rotateY",
+      },
+    });
   });
 });
