+++ packages/css-data/src/property-parsers/index.ts
@@ -1,5 +1,4 @@
 export * from "./transition";
 export * from "./transition-property-extractor";
 export * from "./shadow-properties-extractor";
-export * from "./rotate-properties-extractor";
 export * from "./skew-properties-extractor";
